rendering list in admin home
