+++ client/src/components/index.js
@@ -7,3 +7,4 @@ export { default as MediatorList } from './MediatorList';
 export {default as MediatorCasesShow} from './MediatorCasesShow';
 export {default as CaseDetails} from './CaseDetails';
 export { default as UserCaseList } from './UserCaseList';
+export {default as AdminMediatorRequestList} from './AdminMediatorRequestList';
